Fixing some issues with the how the compiled c++ code connects with the package
Required generating a package.R file taht appears to auto-generate an Rd file for the package. Also fixed an issue where compiling fastdist locally would work, but compiling it as part of the package led to a std::bad_alloc error.  Removing the #define ARMA_64BIT_WORD 1 seems to have solved that problem.