Removing sigma_val as an option
Based on our discussion in issue #3 , a very small sigma leads to numerical issues when calculating eigenvalues/vectors. For the continuous case, we might as well set a large (>0) value of sigma, and for the discrete case we shouldn't use sigma at all. In neither case should it be a user-facing numerical control.

I also made the calculate_eigenfunctions function now include an explicit discrete = TRUE/FALSE argument, so it is easier to specify in code.